break out VolumeSerializer
